buildbot/requirements.txt
buildbot
buildbot-console-view
buildbot-prometheus
buildbot-waterfall-view
buildbot-worker
buildbot-www
libarchive
psycopg2
pynacl
requests
txrequests
